Created new home page with persistent nav bar and correct navigation.
