+++ lib/create_project_and_teams.dart
@@ -82,8 +82,6 @@ class _CreateProjectAndTeamsPageState extends State<CreateProjectAndTeamsPage> {
             ),
           ),
         ),
-
-        bottomNavigationBar: const BottomFloatingNavBar(),
       ),
     );
   }
