+++ lib/main.dart
@@ -10,6 +10,7 @@ import 'teams_and_invites_page.dart';
 import 'login_screen.dart';
 import 'signup_screen.dart';
 import 'home_screen.dart';
+import 'new_home_page.dart';
 
 void main() {
   runApp(const MyApp());
@@ -40,6 +41,7 @@ class MyApp extends StatelessWidget {
         '/login': (context) => const LoginScreen(),
         '/signup': (context) => const SignUpScreen(),
         '/home': (context) => const HomeScreen(),
+        '/new_home': (context) => const BottomFloatingNavBar(),
       },
     );
   }
@@ -135,6 +137,12 @@ class _HomePageStates extends State<HomePage> {
                 route: '/home',
                 name: 'Home',
                 version: 1,
+              ),
+              buildTempButton(
+                context: context,
+                route: '/new_home',
+                name: 'New Home Page',
+                version: 0,
               ),
             ],
           ),
